+++ packages/core/src/storage/kv-store/index.ts
@@ -1,8 +1,7 @@
 import { fs, path } from "@llamaindex/env";
-import { IndexStruct, jsonToIndexStruct } from "../../data-structs";
-import { DEFAULT_COLLECTION, DEFAULT_NAMESPACE } from "../../global";
+
+import { DEFAULT_COLLECTION } from "../../global";
 import type { StoredValue } from "../../schema";
-import { BaseIndexStore } from "../index-store";
 
 async function exists(path: string): Promise<boolean> {
   try {
@@ -124,47 +123,5 @@ export class SimpleKVStore extends BaseKVStore {
 
   static fromDict(saveDict: DataType): SimpleKVStore {
     return new SimpleKVStore(saveDict);
-  }
-}
-
-export class KVIndexStore extends BaseIndexStore {
-  private _kvStore: BaseKVStore;
-  private _collection: string;
-
-  constructor(kvStore: BaseKVStore, namespace: string = DEFAULT_NAMESPACE) {
-    super();
-    this._kvStore = kvStore;
-    this._collection = `${namespace}/data`;
-  }
-
-  async addIndexStruct(indexStruct: IndexStruct): Promise<void> {
-    const key = indexStruct.indexId;
-    const data = indexStruct.toJson();
-    await this._kvStore.put(key, data, this._collection);
-  }
-
-  async deleteIndexStruct(key: string): Promise<void> {
-    await this._kvStore.delete(key, this._collection);
-  }
-
-  async getIndexStruct(structId?: string): Promise<IndexStruct | undefined> {
-    if (!structId) {
-      const structs = await this.getIndexStructs();
-      if (structs.length !== 1) {
-        throw new Error("More than one index struct found");
-      }
-      return structs[0];
-    } else {
-      const json = await this._kvStore.get(structId, this._collection);
-      if (json == null) {
-        return;
-      }
-      return jsonToIndexStruct(json);
-    }
-  }
-
-  async getIndexStructs(): Promise<IndexStruct[]> {
-    const jsons = await this._kvStore.getAll(this._collection);
-    return Object.values(jsons).map((json) => jsonToIndexStruct(json));
   }
 }
